continue training function with checkpoints implemented. Logging system for template implemented

+++ SR/trainer/base_trainer.py
@@ -1,13 +1,16 @@
+from multiprocessing import Value
 import os
 import sys
 import torch
 import pathlib
 import shutil
+import logging
 from numpy import inf
 from tqdm import tqdm
 from abc import abstractmethod
 import matplotlib.pyplot as plt
 import test_all
+from utils import util
 
 
 class BaseTrainer:
@@ -24,16 +27,35 @@ class BaseTrainer:
         self.epochs = config['epochs']
         self.save_period = config['save_period']
         self.start_epoch = config['start_epoch']
+        if self.start_epoch == 0:
+            raise ValueError("start_epoch must start from at least 1")
         self.progress_bar = None
         self.train_loss = []
         self.valid_loss = []
         self.checkpoint_dir = pathlib.Path(config['checkpoint_dir'])
-        if self.checkpoint_dir.exists():
-            shutil.rmtree(self.checkpoint_dir)
+        self.valid_loss_plot_name, self.train_loss_plot_name = util.get_loss_plot_names(
+            self.checkpoint_dir)
         self.model_weights_dir = self.checkpoint_dir / 'weights'
         self.valid_results = self.checkpoint_dir / 'validation'
+        self.log_path = self.checkpoint_dir / 'log.log'
+        if self.start_epoch <= 1 and self.checkpoint_dir.exists():
+            # not training from a checkpoint
+            shutil.rmtree(self.checkpoint_dir)
+        else:
+            # load weights
+            logging.info(
+                f"'start_epoch' is not 1, looking for epoch{self.start_epoch}.pth")
+            weights_files = os.listdir(self.model_weights_dir)
+            if f'epoch{self.start_epoch}.pth' in weights_files:
+                logging.info(
+                    f"epoch{self.start_epoch}.pth found, load model weights")
+                self.model.load_state_dict(torch.load(
+                    self.model_weights_dir/f'epoch{self.start_epoch}.pth'))
+            else:
+                raise ValueError(
+                    f"Weight file not found, the start epoch is {self.start_epoch}, epoch{self.start_epoch}.pth doesn't exist in {self.model_weights_dir}")
         for path in [self.checkpoint_dir, self.model_weights_dir, self.valid_results]:
-            path.mkdir(parents=True, exist_ok=False)
+            path.mkdir(parents=True, exist_ok=True)
 
     @abstractmethod
     def _train_epoch(self, epoch):
@@ -44,7 +66,7 @@ class BaseTrainer:
         raise NotImplementedError
 
     def train(self):
-        with tqdm(total=len(self.train_dataset) * self.epochs) as progress_bar:
+        with tqdm(total=len(self.train_dataset) * (self.epochs - self.start_epoch + 1)) as progress_bar:
             # with tqdm(range(self.start_epoch, self.epochs + 1), total=self.epochs, file=sys.stdout) as progress_bar:
             self.progress_bar = progress_bar
             for epoch in range(self.start_epoch, self.epochs + 1):
@@ -57,26 +79,31 @@ class BaseTrainer:
         self.progress_bar.close()
 
     def _save_checkpoint(self, epoch):
+
         torch.save(self.model.state_dict(), os.path.join(
             self.model_weights_dir, 'epoch{}.pth'.format(epoch)))
         # training loss plot
         if len(self.train_loss) != 0:
+            logging.debug("plot training loss")
             plt.figure()
-            plt.plot(list(range(1, len(self.train_loss) + 1)), self.train_loss)
+            plt.plot(list(range(self.start_epoch, self.start_epoch +
+                                len(self.train_loss))), self.train_loss)
             plt.xlabel("epoch")
             plt.ylabel("loss")
             plt.title("Training Loss")
-            plt.savefig(self.checkpoint_dir / 'train_loss.png')
+            plt.savefig(self.train_loss_plot_name)
             plt.close()
         else:
-            print("error: no training loss")
+            logging.error("error: no training loss")
         if len(self.valid_loss) != 0:
+            logging.debug("plot validation loss")
             plt.figure()
-            plt.plot(list(range(1, len(self.valid_loss) + 1)), self.valid_loss)
+            plt.plot(list(range(self.start_epoch, self.start_epoch +
+                                len(self.valid_loss))), self.valid_loss)
             plt.xlabel("epoch")
             plt.ylabel("loss")
             plt.title("Validation Loss")
-            plt.savefig(self.checkpoint_dir / 'valid_loss.png')
+            plt.savefig(self.valid_loss_plot_name)
             plt.close()
         else:
-            print("error: no validation loss")
+            logging.error("error: no validation loss")
